backend/requirements.txt
# FastAPI and server
fastapi==0.115.7
uvicorn[standard]==0.34.0
python-multipart==0.0.6

# Pydantic for settings and validation
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.0

# Async HTTP client for HuggingFace API
httpx==0.28.1

# Image processing
Pillow==10.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0  # passlib 1.7.4 not compatible with bcrypt 5.x

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0

# Background tasks
apscheduler==3.10.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Code quality
black==24.10.0
flake8==7.1.1
mypy==1.14.1
